requirements.txt
discord.py
jishaku

asyncpg>=0.5.4
uvloop
